feat: working darwin version

+++ default.nix
@@ -1,10 +1,12 @@
 {
+  stdenv,
   lib,
   buildGoModule,
   buildNpmPackage,
   fetchFromGitHub,
   git,
   electron_36,
+  rsync,
 }:
 let
   electron = electron_36;
@@ -60,23 +62,24 @@ buildNpmPackage {
 
   nativeBuildInputs = [
     electron
+    rsync # see scripts/after-sync.js
   ];
 
   env.ELECTRON_SKIP_BINARY_DOWNLOAD = "1";
 
   postPatch = ''
     substituteInPlace package.json --replace-fail '"beforeBuild": "./scripts/build-backend.js",' ""
-    substituteInPlace package.json --replace-fail '"../backend/headlamp-server"' '"${backend}/bin/headlamp-server"'
-    substituteInPlace package.json --replace-fail '"../frontend/build"' '"${frontend}/build"'
+    substituteInPlace package.json --replace-fail '../backend/headlamp-server' '${backend}/bin/headlamp-server'
+    substituteInPlace package.json  --replace-fail '../frontend/build' '${frontend}/build'
 
     # electron complains otherwise
     substituteInPlace package.json --replace-fail '"asar": false,' ""
-
+  ''
+  + lib.optionalString stdenv.hostPlatform.isLinux ''
     # https://github.com/electron/electron/issues/31121
     substituteInPlace electron/main.ts \
       --replace-fail "process.resourcesPath" "'$out/share/headlamp/resources'"
   '';
-
   buildPhase = ''
     runHook preBuild
 
@@ -84,10 +87,25 @@ buildNpmPackage {
     npm run copy-plugins
     npm run compile-electron
     npm run prod-deps
-    npm exec electron-builder -- --dir --publish never \
-      -c.electronDist=${electron.dist} \
-      -c.electronVersion=${electron.version} \
-      -c.npmRebuild=false
+    ${
+      if stdenv.hostPlatform.isDarwin then
+        ''
+          cp -r ${electron.dist}/Electron.app ./
+          find ./Electron.app -name 'Info.plist' | xargs -d '\n' chmod +rw
+
+          npm exec electron-builder -- --dir --publish never \
+            -c.electronDist=./ \
+            -c.electronVersion=${electron.version} \
+            -c.npmRebuild=false
+        ''
+      else
+        ''
+          npm exec electron-builder -- --dir --publish never \
+            -c.electronDist=${electron.dist} \
+            -c.electronVersion=${electron.version} \
+            -c.npmRebuild=false
+        ''
+    }
 
     runHook postBuild
   '';
@@ -95,17 +113,28 @@ buildNpmPackage {
   installPhase = ''
     runHook preInstall
 
-    mkdir -p $out/share/headlamp
-    pushd dist/linux-unpacked
-    # TODO: some of these point into /build, not sure how to avoid installing these.
-    rm -rf resources/app/node_modules/*/node_modules/.bin
-    cp -r locales resources{,.pak} $out/share/headlamp
-    popd
+    ${
+      if stdenv.hostPlatform.isDarwin then
+        ''
+          mkdir -p $out/Applications
+          cp -r dist/*/Headlamp.app $out/Applications/
+        ''
+      else
+        ''
+          mkdir -p $out/share/headlamp
+          pushd dist/linux-unpacked
+          # TODO: some of these point into /build, not sure how to avoid installing these.
+          rm -rf resources/app/node_modules/*/node_modules/.bin
+          cp -r locales resources{,.pak} $out/share/headlamp
+          popd
 
-    makeWrapper ${lib.getExe electron} $out/bin/headlamp \
-      --add-flags $out/share/headlamp/resources/app.asar \
-      --add-flags "\''${NIXOS_OZONE_WL:+\''${WAYLAND_DISPLAY:+--ozone-platform-hint=auto --enable-features=WaylandWindowDecorations --enable-wayland-ime=true}}" \
-      --inherit-argv0
+          makeWrapper ${lib.getExe electron} $out/bin/headlamp \
+            --add-flags $out/share/headlamp/resources/app.asar \
+            --add-flags "\''${NIXOS_OZONE_WL:+\''${WAYLAND_DISPLAY:+--ozone-platform-hint=auto --enable-features=WaylandWindowDecorations --enable-wayland-ime=true}}" \
+            --inherit-argv0
+        ''
+    }
+
   '';
 
   meta = with lib; {
